requirements.txt
absl-py==2.1.0
ace==0.3.3
alpha-transform==0.0.1
appdirs==1.4.4
astropy==5.2.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.16.1
fonttools==4.53.1
fsspec==2025.2.0
gecatsim==1.3.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-auth-oauthlib==1.0.0
grpcio==1.66.1
h5py==3.11.0
healpy==1.16.5
idna==3.8
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
Jinja2==3.1.5
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numexpr==2.8.6
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pathtools==0.1.2
pillow==10.4.0
plotly==5.24.1
protobuf==4.25.4
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyerfa==2.0.0.3
pyFFTW==0.13.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==2.13.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
tifffile==2023.7.10
torch==2.4.1
torchaudio==0.7.2
torchmetrics==1.4.2
torchvision==0.8.2+cu110
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
wandb==0.15.11
Werkzeug==3.0.4
zipp==3.20.1
